+++ src/main/java/com/financefit/financeFit/dtos/CreateDespesaDTO.java
@@ -15,7 +15,7 @@ public class CreateDespesaDTO {
     private String descricao;
 
     @NotNull(message = "ID do usuário é obrigatório")
-    private Integer idUsuario;
+    private Long idUsuario;
 
     @NotNull(message = "ID da categoria é obrigatório")
     private Long idCategoria;
@@ -25,7 +25,7 @@ public class CreateDespesaDTO {
     public CreateDespesaDTO() {
     }
 
-    public CreateDespesaDTO(BigDecimal valor, LocalDate data, String descricao, Integer idUsuario, Long idCategoria, TipoTransacao tipo) {
+    public CreateDespesaDTO(BigDecimal valor, LocalDate data, String descricao, Long idUsuario, Long idCategoria, TipoTransacao tipo) {
         this.valor = valor;
         this.data = data;
         this.descricao = descricao;
@@ -58,11 +58,11 @@ public class CreateDespesaDTO {
         this.descricao = descricao;
     }
 
-    public Integer getIdUsuario() {
+    public Long getIdUsuario() {
         return idUsuario;
     }
 
-    public void setIdUsuario(Integer idUsuario) {
+    public void setIdUsuario(Long idUsuario) {
         this.idUsuario = idUsuario;
     }
