+++ src/main/java/com/financefit/financeFit/entities/Despesa.java
@@ -11,7 +11,7 @@ public class Despesa {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "despesa_id")
-    private int id;
+    private Long id;
 
     @Column(name = "valor", nullable = false, precision = 10, scale = 2)
     private BigDecimal valor;
@@ -37,7 +37,7 @@ public class Despesa {
     public Despesa() {
     }
 
-    public Despesa(int id, BigDecimal valor, LocalDate data, String descricao, Usuario usuario, Categoria categoria) {
+    public Despesa(Long id, BigDecimal valor, LocalDate data, String descricao, Usuario usuario, Categoria categoria) {
         this.id = id;
         this.valor = valor;
         this.data = data;
@@ -47,11 +47,11 @@ public class Despesa {
         this.tipo = TipoTransacao.DESPESA; // Garantir que o tipo seja DESPESA
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
