+++ src/main/java/com/financefit/financeFit/repositories/DespesaRepository.java
@@ -10,14 +10,14 @@ import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
-public interface DespesaRepository extends JpaRepository<Despesa, Integer> {
+public interface DespesaRepository extends JpaRepository<Despesa, Long> {
     @Query("SELECT SUM(d.valor) FROM Despesa d WHERE d.usuario.userId = :idUsuario AND MONTH(d.data) = :mes AND YEAR(d.data) = :ano")
     BigDecimal calcularTotalGastoNoMes(
-            @Param("idUsuario") int idUsuario,
+            @Param("idUsuario") Long idUsuario,
             @Param("mes") int mes,
             @Param("ano") int ano
     );
 
     @Query("SELECT d FROM Despesa d JOIN FETCH d.usuario JOIN FETCH d.categoria WHERE d.usuario.userId = :idUsuario")
-    List<Despesa> findByUsuarioUserId(@Param("idUsuario") int idUsuario);
+    List<Despesa> findByUsuarioUserId(@Param("idUsuario") Long idUsuario);
 }
